requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
idna==3.4
psycopg==3.1.7
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.8.22
geopandas==0.12.2
idna==3.4
munch==2.5.0
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
shapely==2.0.0
six==1.16.0
django-cors-headers==3.13.0
